fixing more configs.. adding sudo to tasks that needed it
some more configs needed to be fixed...

also i had to add sudo to tasks to get them to work..

it feel sleek the whole deploy/ubuntu user permissions and user accounts should be reviewed..

I'm not sure ansible scripts were idempotent at this point. so bear this in mind.

Signed-off-by: Ryan Medlin <ryan.w.medlin@gmail.com>