IMPORTANT: flattened security for apps to only use ubuntu user
I spent a few hours on this and there is some very strange stuff going on in trying to get the deploy user working.

I either do not understand it and its hard to get information on what users and keys are being used where.

to keep making forward progress i flattened to the ubuntu user.

we will need to refit the use of the deploy user and get that integrated back in properly as a sysadmin would normally want.

Signed-off-by: Ryan Medlin <ryan.w.medlin@gmail.com>